Changed all entity's positions to vectors and floats.

+++ ProgramGeist2/src/codeBlock/CBlock.java
@@ -8,7 +8,7 @@ public class CBlock extends CodeBlock {
 	
 	protected ArrayList<CodeBlock> stack = new ArrayList<CodeBlock>();
 	
-	public CBlock(double x, double y, EntityWorld world) {
+	public CBlock(float x, float y, EntityWorld world) {
 		super(x, y, world);
 	}
 	public CBlock(CodeBlock downBlock, EntityWorld world) {
